Use usize instead of u8 for error count

+++ src/batsign.rs
@@ -59,7 +59,7 @@ pub fn send_batsign_notification(
         println!("Batsigns sent; HTTP statuses: {:?}", statuses);
     }
 
-    let mut num_errors: u8 = 0;
+    let mut num_errors: usize = 0;
 
     for status in statuses {
         if !status.is_success() {
